--- src/main/java/se/sundsvall/teamssender/service/TeamsMessageService.java
@@ -1,12 +1,0 @@
-package se.sundsvall.teamssender.service;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Service;
-
-@Service
-public class TeamsMessageService {
-
-	public ResponseEntity<String> sendTeamsMessage() {
-		return ResponseEntity.ok("Sent successfully");
-	}
-}
